Move auth.js to middlewares folder
